improve utility function to be a float64 that takes into account the point values of both sides AND the mobility of both sides

+++ go/src/chess/structs.go
@@ -6,7 +6,7 @@ type ChessNode struct {
 	prev_move     Move
 	legal_moves   []Move
 	depth         int
-	utility_value int
+	utility_value float64
 }
 
 type Move struct {
